ref: inputs in CreateWord are being seperated into individual components
