+++ src/_pages/Dashboard/CreateWordForm/Inputs/WordPrefixInput.tsx
@@ -1,0 +1,26 @@
+import { WordForm } from '@/types'
+import { Input } from '@nextui-org/react'
+import React from 'react'
+import { Control, Controller } from 'react-hook-form'
+
+export default function WordPrefixInput({
+    control
+}: {
+    control: Control<WordForm>
+}) {
+    return (
+        <Controller
+            control={control}
+            name="prefix"
+            render={({ field }) => (
+                <Input
+                    {...field}
+                    label="Prefix"
+                    color="primary"
+                    variant="underlined"
+                    description="Prefix is optional"
+                />
+            )}
+        />
+    )
+}
